add general perl and python scripts
The scripts for basic and general operations used in this analysis are included in the `bin` directory in this commit. ScriptManager JAR file should go in the same directory but I will allow the user to add it so this is also added to the .gitignore file with this commit.