add RNA normalization to preprocessing
This commit includes the lines for running two different kinds of normalization on the aligned RNA BAM files, replacing the "Skip Pol-A RNAseq" print placeholder. In addition, some paths were fixed and all moved to be defined together near the top. Thirdly, code was added so that the ScriptManager JAR file is copied at the beginning of each run to use and then removed at the end.